% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainsplot.R
\name{gainsplot}
\alias{gainsplot}
\title{Plotting Gains Curves}
\usage{
gainsplot(outcome.var, ...)
}
\arguments{
\item{outcome.var, }{a binary outcome variable to be assessed by the scores}

\item{..., }{one or more scores to assess the outcome variable}
}
\description{
This function plots the gains curve and returns the area under the gains curve (AUGC) for one or more scores.
}
\examples{
gainsplot(-lean$rec, lean$freq, lean$mon, outcome.var=lean$offer)
}
