% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardepplot.R
\name{pardepplot}
\alias{pardepplot}
\title{Plotting Partial Dependence}
\usage{
pardepplot(object, pred.var, data, ylim = NULL, hline = NULL)
}
\arguments{
\item{object:}{a model of class lm, glm, nnet, or ranger.}

\item{pred.var:}{a character or a vector of two characters giving the predictor variables for which a partial dependence plot is sought; predictor variables must be of type numeric or factor.}

\item{data:}{a dataframe supplying the variables for plotting; typically the dataframe used to estimate the model.}

\item{ylim:}{optional vector of length two specifying the y-axis limits.}

\item{hline:}{optimal number specifying the location of a horizontal line.}
}
\description{
This function plots partial dependence. It relies on the pdp package partial function.
}
\examples{
pardepplot(lr, pred.var="art", data=bbb)
}
