% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift.R
\name{QiniBarPlot}
\alias{QiniBarPlot}
\title{Plotting Qini Bar Plots}
\usage{
QiniBarPlot(..., modelnames = NULL)
}
\arguments{
\item{..., }{one or more Qini tables}

\item{modelnames}{= c("model1", "model2") (optional)}
}
\description{
This function plots the Qini bar plot for one or more models.
}
\examples{
QiniBarPlot(PerfTable_uplift, PerfTable_propensity, modelnames = c("Logit Uplift", "Logit Propensity"))
}
