% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift.R
\name{QiniCurve}
\alias{QiniCurve}
\title{Plotting Qini Curves}
\usage{
QiniCurve(..., modelnames = NULL)
}
\arguments{
\item{..., }{one or more Qini tables}

\item{modelnames}{= c("model1", "model2") (optional)}
}
\description{
This function plots the Qini curve for one or more models.
}
\examples{
QiniCurve2(PerfTable_uplift, PerfTable_propensity, modelnames = c("Logit Uplift", "Logit Propensity"))
}
