% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimpplot.R
\name{varimpplot}
\alias{varimpplot}
\title{Plotting Variable Importance}
\usage{
varimpplot(object, target, data)
}
\arguments{
\item{object:}{a model of class lm, glm, nnet, or ranger.}

\item{data:}{a dataframe on which to evaluate variable importance; typically a test dataframe.}
}
\description{
This function plots variable importance. It telies on the vip package vi function.
}
\examples{
varimpplot(lr, target="buyer", data=bbb)
}
