% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift.R
\name{QiniBarPlot}
\alias{QiniBarPlot}
\title{Plotting Qini Bar Plots}
\usage{
QiniBarPlot(..., score.names = NULL)
}
\arguments{
\item{..., }{one or more Qini tables}

\item{score.names}{= c("score1", "score2") (optional)}
}
\description{
This function plots the Qini bar plot for one or more scores.
}
\examples{
QiniBarPlot(PerfTable_uplift, PerfTable_propensity, score.names = c("LR Uplift", "LR Propensity"))
}
