% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift.R
\name{QiniCurve}
\alias{QiniCurve}
\title{Plotting Qini Curves}
\usage{
QiniCurve(..., score.names = NULL)
}
\arguments{
\item{..., }{one or more Qini tables}

\item{score.names}{= c("score1", "score2") (optional)}
}
\description{
This function plots the Qini curve for one or more scores.
}
\examples{
QiniCurve2(PerfTable_uplift, PerfTable_propensity, score.names = c("LR Uplift", "LR Propensity"))
}
