% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift.R
\name{QiniTable}
\alias{QiniTable}
\title{Computing Qini Tables}
\usage{
QiniTable(data, treat, outcome, prediction, nb.group = 10)
}
\arguments{
\item{data:}{a data frame containing the treatment, the outcome and the predictors.}

\item{treat:}{name of a binary (numeric) vector representing the treatment assignment (coded as 0/1)}

\item{outcome:}{name of a binary response (numeric) vector (coded as 0/1)}

\item{prediction:}{a score to sort the observations from highest to lowest uplift or propensity.}

\item{nb.group:}{(optional, default = 10)}
}
\description{
This function computes the Qini table used to evaluate the performance of an uplift model.
}
\examples{
QiniTable(
 expdata_stacked.test,
 treat = "ad",
 outcome = "converted",
 prediction = "score_logitFit_treat",
 nb.group = 20
)
}
