% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_kelloggmkgt482}
\alias{update_kelloggmkgt482}
\title{Updating kelloggmktg482}
\usage{
update_kelloggmkgt482()
}
\description{
This function updates the kelloggmkgt482 package
}
\examples{
\dontrun{
update_kelloggmkgt482()
}
}
