% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_kelloggmkgt482}
\alias{update_kelloggmkgt482}
\title{A function to update the kelloggmkgt482 package}
\usage{
update_kelloggmkgt482()
}
\description{
This function allows you to easily update the mkgt482 package from github
}
\examples{
update_kelloggmkgt482()
}
\keyword{update}
