% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimpplot.R
\name{varimpplot}
\alias{varimpplot}
\title{A function to plot a variable importance plot.}
\usage{
varimpplot(object, target, data = NULL)
}
\arguments{
\item{object:}{a model of class lm, glm, nnet, or ranger.}

\item{data:}{a dataframe on which to evaluate variable importance; optional for models of all classes except ranger; typically a test dataframe.}
}
\description{
Relies on the vip package vi function.
}
\examples{
varimpplot(lr, target="buyer")
}
